Clean up organization for OpRunner.
