+++ test/CMakeLists.txt
@@ -3,7 +3,7 @@ set_source_path(TEST_BASE_HDRS
   test_utils.hpp
   test_utils_mpi.hpp
   op_dispatcher.hpp
-  op_runner_impl.hpp
+  op_runner.hpp
   hang_watchdog.hpp)
 if (AL_HAS_CUDA)
   set_source_path(TEST_CUDA_HDRS cuda_vector.hpp)
