+++ test/op_runner.hpp
@@ -28,6 +28,150 @@
 #pragma once
 
 #include "Al.hpp"
+#include "aluminum/traits/traits.hpp"
+#include "algo_support.hpp"
+#include "test_utils.hpp"
+
+/** Pass options that may be used by an operator. */
+template <typename Backend>
+struct OpOptions {
+  bool inplace = false;
+  bool nonblocking = false;
+  int root = 0;
+  int src = -1;
+  int dst = -1;
+  std::vector<size_t> send_counts = {};
+  std::vector<size_t> send_displs = {};
+  std::vector<size_t> recv_counts = {};
+  std::vector<size_t> recv_displs = {};
+  Al::ReductionOperator reduction_op = Al::ReductionOperator::sum;
+  typename Backend::req_type req = Backend::null_req;
+  AlgorithmOptions<Backend> algos;
+};
+
+/** Abstract base class for running an operator. */
+template <typename Backend, typename T, typename Child>
+class OpRunnerBase {
+public:
+  OpRunnerBase(OpOptions<Backend> &options_) :
+    options(options_) {}
+  ~OpRunnerBase() {}
+
+  const OpOptions<Backend>& get_options() const { return options; }
+  OpOptions<Backend>& get_options() { return options; }
+  std::string get_name() const {
+    return static_cast<const Child*>(this)->get_name_impl();
+  };
+
+  void run(typename VectorType<T, Backend>::type& input,
+           typename VectorType<T, Backend>::type& output,
+           typename Backend::comm_type& comm) {
+    static_cast<Child*>(this)->run_impl_int(input, output, comm);
+  }
+  void run_mpi(std::vector<T>& input,
+               std::vector<T>& output,
+               typename Backend::comm_type& comm) {
+    static_cast<Child*>(this)->run_mpi_impl_int(input, output, comm);
+  }
+
+  size_t get_input_size(size_t base_size,
+                        typename Backend::comm_type& comm) {
+    return static_cast<Child*>(this)->get_input_size_int_impl(base_size, comm);
+  }
+  size_t get_output_size(size_t base_size,
+                         typename Backend::comm_type& comm) {
+    return static_cast<Child*>(this)->get_output_size_int_impl(base_size, comm);
+  }
+
+protected:
+  void inplace_nb_dispatch(std::function<void()> nip_b,
+                           std::function<void()> ip_b,
+                           std::function<void()> nip_nb,
+                           std::function<void()> ip_nb) {
+    if (get_options().inplace) {
+      if (get_options().nonblocking) {
+        ip_nb();
+      } else {
+        ip_b();
+      }
+    } else {
+      if (get_options().nonblocking) {
+        nip_nb();
+      } else {
+        nip_b();
+      }
+    }
+  }
+
+  void* buf_or_inplace(T* buf) {
+    return get_options().inplace ? MPI_IN_PLACE : buf;
+  }
+
+private:
+  OpOptions<Backend>& options;
+};
+
+/**
+ * Intermediate ABC providing some common functions.
+ */
+template <Al::AlOperation Op, typename Backend, typename T, typename Child>
+class OpRunnerShim : public OpRunnerBase<Backend, T, OpRunnerShim<Op, Backend, T, Child>> {
+public:
+  using OpRunnerBase<Backend, T, OpRunnerShim<Op, Backend, T, Child>>::OpRunnerBase;
+
+  std::string get_name_impl() const { return Al::AlOperationName<Op>; }
+
+  template <Al::AlOperation Op2 = Op,
+            std::enable_if_t<Al::IsOpSupported<Op2, Backend>::value, bool> = true>
+  void run_impl_int(typename VectorType<T, Backend>::type& input,
+                    typename VectorType<T, Backend>::type& output,
+                    typename Backend::comm_type& comm) {
+    static_cast<Child*>(this)->run_impl(input, output, comm);
+  }
+  template <Al::AlOperation Op2 = Op,
+            std::enable_if_t<!Al::IsOpSupported<Op2, Backend>::value, bool> = true>
+  void run_impl_int(typename VectorType<T, Backend>::type&,
+                    typename VectorType<T, Backend>::type&,
+                    typename Backend::comm_type&) {
+    std::cerr << Al::AlOperationName<Op> << " not supported by backend" << std::endl;
+    std::abort();
+  }
+
+  template <typename T2 = T,
+            std::enable_if_t<Al::IsTypeSupportedByMPI<T2>::value, bool> = true>
+  void run_mpi_impl_int(std::vector<T>& input,
+                        std::vector<T>& output,
+                        typename Backend::comm_type& comm) {
+    static_cast<Child*>(this)->run_mpi_impl(input, output, comm);
+  }
+  template <typename T2 = T,
+            std::enable_if_t<!Al::IsTypeSupportedByMPI<T2>::value, bool> = true>
+  void run_mpi_impl_int(std::vector<T>&, std::vector<T>&,
+                        typename Backend::comm_type&) {
+    std::cerr << "Type not supported by MPI" << std::endl;
+    std::abort();
+  }
+
+  size_t get_input_size_int_impl(size_t base_size,
+                                 typename Backend::comm_type& comm) {
+    return static_cast<Child*>(this)->get_input_size_impl(base_size, comm);
+  }
+  size_t get_output_size_int_impl(size_t base_size,
+                                  typename Backend::comm_type& comm) {
+    return static_cast<Child*>(this)->get_output_size_impl(base_size, comm);
+  }
+
+};
+
+/**
+ * Main OpRunner class, handles calls to Aluminum/MPI for an operator.
+ *
+ * Should be specialized for specific operators.
+ */
+template <Al::AlOperation Op, typename Backend, typename T>
+class OpRunner : public OpRunnerShim<Op, Backend, T, OpRunner<Op, Backend, T>> {};
+
+// Specific implementations are below:
 
 template <typename Backend, typename T>
 class OpRunner<Al::AlOperation::allgather, Backend, T> :
@@ -55,7 +199,7 @@ public:
   }
 
   template <typename T2 = T,
-            std::enable_if_t<IsTypeSupportedByMPI<T2>::value, bool> = true>
+            std::enable_if_t<Al::IsTypeSupportedByMPI<T2>::value, bool> = true>
   void run_mpi_impl(std::vector<T>& input,
                     std::vector<T>& output,
                     typename Backend::comm_type& comm) {
@@ -107,7 +251,7 @@ public:
   }
 
   template <typename T2 = T,
-            std::enable_if_t<IsTypeSupportedByMPI<T2>::value, bool> = true>
+            std::enable_if_t<Al::IsTypeSupportedByMPI<T2>::value, bool> = true>
   void run_mpi_impl(std::vector<T>& input,
                     std::vector<T>& output,
                     typename Backend::comm_type& comm) {
@@ -161,7 +305,7 @@ public:
   }
 
   template <typename T2 = T,
-            std::enable_if_t<IsTypeSupportedByMPI<T2>::value, bool> = true>
+            std::enable_if_t<Al::IsTypeSupportedByMPI<T2>::value, bool> = true>
   void run_mpi_impl(std::vector<T>& input,
                     std::vector<T>& output,
                     typename Backend::comm_type& comm) {
@@ -214,7 +358,7 @@ public:
   }
 
   template <typename T2 = T,
-            std::enable_if_t<IsTypeSupportedByMPI<T2>::value, bool> = true>
+            std::enable_if_t<Al::IsTypeSupportedByMPI<T2>::value, bool> = true>
   void run_mpi_impl(std::vector<T>& input,
                     std::vector<T>& output,
                     typename Backend::comm_type& comm) {
@@ -270,7 +414,7 @@ public:
   }
 
   template <typename T2 = T,
-            std::enable_if_t<IsTypeSupportedByMPI<T2>::value, bool> = true>
+            std::enable_if_t<Al::IsTypeSupportedByMPI<T2>::value, bool> = true>
   void run_mpi_impl(std::vector<T>& input,
                     std::vector<T>& output,
                     typename Backend::comm_type& comm) {
@@ -335,7 +479,7 @@ public:
   }
 
   template <typename T2 = T,
-            std::enable_if_t<IsTypeSupportedByMPI<T2>::value, bool> = true>
+            std::enable_if_t<Al::IsTypeSupportedByMPI<T2>::value, bool> = true>
   void run_mpi_impl(std::vector<T>& /*input*/,
                     std::vector<T>& /*output*/,
                     typename Backend::comm_type& comm) {
@@ -381,7 +525,7 @@ public:
   }
 
   template <typename T2 = T,
-            std::enable_if_t<IsTypeSupportedByMPI<T2>::value, bool> = true>
+            std::enable_if_t<Al::IsTypeSupportedByMPI<T2>::value, bool> = true>
   void run_mpi_impl(std::vector<T>& /*input*/,
                     std::vector<T>& output,
                     typename Backend::comm_type& comm) {
@@ -432,7 +576,7 @@ public:
   }
 
   template <typename T2 = T,
-            std::enable_if_t<IsTypeSupportedByMPI<T2>::value, bool> = true>
+            std::enable_if_t<Al::IsTypeSupportedByMPI<T2>::value, bool> = true>
   void run_mpi_impl(std::vector<T>& input,
                     std::vector<T>& output,
                     typename Backend::comm_type& comm) {
@@ -500,7 +644,7 @@ public:
   }
 
   template <typename T2 = T,
-            std::enable_if_t<IsTypeSupportedByMPI<T2>::value, bool> = true>
+            std::enable_if_t<Al::IsTypeSupportedByMPI<T2>::value, bool> = true>
   void run_mpi_impl(std::vector<T>& input,
                     std::vector<T>& output,
                     typename Backend::comm_type& comm) {
@@ -570,7 +714,7 @@ public:
   }
 
   template <typename T2 = T,
-            std::enable_if_t<IsTypeSupportedByMPI<T2>::value, bool> = true>
+            std::enable_if_t<Al::IsTypeSupportedByMPI<T2>::value, bool> = true>
   void run_mpi_impl(std::vector<T>& input,
                     std::vector<T>& output,
                     typename Backend::comm_type& comm) {
@@ -637,7 +781,7 @@ public:
   }
 
   template <typename T2 = T,
-            std::enable_if_t<IsTypeSupportedByMPI<T2>::value, bool> = true>
+            std::enable_if_t<Al::IsTypeSupportedByMPI<T2>::value, bool> = true>
   void run_mpi_impl(std::vector<T>& input,
                     std::vector<T>& output,
                     typename Backend::comm_type& comm) {
@@ -693,7 +837,7 @@ public:
   }
 
   template <typename T2 = T,
-            std::enable_if_t<IsTypeSupportedByMPI<T2>::value, bool> = true>
+            std::enable_if_t<Al::IsTypeSupportedByMPI<T2>::value, bool> = true>
   void run_mpi_impl(std::vector<T>& input,
                     std::vector<T>& output,
                     typename Backend::comm_type& comm) {
@@ -755,7 +899,7 @@ public:
   }
 
   template <typename T2 = T,
-            std::enable_if_t<IsTypeSupportedByMPI<T2>::value, bool> = true>
+            std::enable_if_t<Al::IsTypeSupportedByMPI<T2>::value, bool> = true>
   void run_mpi_impl(std::vector<T>& input,
                     std::vector<T>& output,
                     typename Backend::comm_type& comm) {
@@ -829,7 +973,7 @@ public:
   }
 
   template <typename T2 = T,
-            std::enable_if_t<IsTypeSupportedByMPI<T2>::value, bool> = true>
+            std::enable_if_t<Al::IsTypeSupportedByMPI<T2>::value, bool> = true>
   void run_mpi_impl(std::vector<T>& input,
                     std::vector<T>& output,
                     typename Backend::comm_type& comm) {
@@ -910,7 +1054,7 @@ public:
   }
 
   template <typename T2 = T,
-            std::enable_if_t<IsTypeSupportedByMPI<T2>::value, bool> = true>
+            std::enable_if_t<Al::IsTypeSupportedByMPI<T2>::value, bool> = true>
   void run_mpi_impl(std::vector<T>& input,
                     std::vector<T>& /*output*/,
                     typename Backend::comm_type& comm) {
@@ -958,7 +1102,7 @@ public:
   }
 
   template <typename T2 = T,
-            std::enable_if_t<IsTypeSupportedByMPI<T2>::value, bool> = true>
+            std::enable_if_t<Al::IsTypeSupportedByMPI<T2>::value, bool> = true>
   void run_mpi_impl(std::vector<T>& /*input*/,
                     std::vector<T>& output,
                     typename Backend::comm_type& comm) {
@@ -1003,7 +1147,7 @@ public:
   }
 
   template <typename T2 = T,
-            std::enable_if_t<IsTypeSupportedByMPI<T2>::value, bool> = true>
+            std::enable_if_t<Al::IsTypeSupportedByMPI<T2>::value, bool> = true>
   void run_mpi_impl(std::vector<T>& input,
                     std::vector<T>& output,
                     typename Backend::comm_type& comm) {
